Chantry creation should create cabals/members and ensure most are of the appropriate faction Fixes #205
