+++ wod/models/characters/mage/__init__.py
@@ -1,5 +1,6 @@
+from .cabal import Cabal
 from .faction import MageFaction
 from .focus import Instrument, Paradigm, Practice
-from .mage import Cabal, Mage, Rote
+from .mage import Mage, Rote
 from .resonance import Resonance, ResRating
 from .rote import Effect
