+++ wod/models/locations/mage/__init__.py
@@ -1,0 +1,3 @@
+from .chantry import Chantry
+from .nodes import Node, NodeMeritFlaw, NodeMeritFlawRating, NodeResonanceRating
+from .sector import Sector
