+++ wod/templates/wod/locations/mage/chantry/detail.html
@@ -9,47 +9,117 @@
                 <h1>{{ object.name }}</h1>
             </div>
             <div class="row">
-            {% if object.parent != None %}
+                {% if object.parent != None %}
                     <div class="col-sm">Parent</div>
                     <a class="col-sm" href="{{ object.get_absolute_url }}">{{ object.parent.name }}</a>
-            {% endif %}
+                {% endif %}
                 <div class="col-sm">Faction</div>
                 <div class="col-sm">{{ factions }}</div>
                 <div class="col-sm">Season</div>
                 <div class="col-sm">{{ object.get_season_display }}</div>
                 <div class="col-sm">Type</div>
-                <div class="col-sm">{{ object.get_chantry_type_display }}</div>    
+                <div class="col-sm">{{ object.get_chantry_type_display }}</div>
             </div>
             <div class="row">
                 <div class="col-sm"></div>
-                <div class="col-sm"></div>
+                <div class="col-sm">Leadership Type</div>
+                <div class="col-sm">{{ object.get_leadership_type_display }}</div>
                 <div class="col-sm">Rank</div>
                 <div class="col-sm">{{ object.rank }}</div>
                 <div class="col-sm"></div>
-                <div class="col-sm"></div>
             </div>
             <div class="row">
-                <div class="col-sm"></div>
-                <div class="col-sm"></div>
-                <div class="col-sm">Reality Zone</div>
-                <div class="col-sm">{{ object.reality_zone }} ({{ object.reality_zone_rating }})</div>
-                <div class="col-sm"></div>
-                <div class="col-sm"></div>
+                <h3 class="col-sm mtas_heading">Leaders</h3>
+            </div>
+            {% for leader in self.leaders.all %}
+                <div class="row">
+                    <a class="col-sm" href="{{ leader.get_absolute_url }}">{{ leader.name }}</a>
+                </div>
+            {% endfor %}
+            <div class="row">
+                <div class="col-sm">Ambassador</div>
+                <div class="col-sm">{{ object.ambassador.name }}</div>
+                <div class="col-sm">Node Tender</div>
+                <div class="col-sm">{{ object.node_tender.name }}</div>
             </div>
             <div class="row">
-                <div class="col-sm">Gauntlet</div>
-                <div class="col-sm">{{ object.gauntlet }}</div>
-                <div class="col-sm">Shroud</div>
-                <div class="col-sm">{{ object.shroud }}</div>
-                <div class="col-sm">Dimension Barrier</div>
-                <div class="col-sm">{{ object.dimension_barrier }}</div>
+                <div class="col-sm">Investigators</div>
+            </div>
+            {% for character in object.investigator.all %}
+                <div class="row">
+                    <div class="col-sm">{{ character.name }}</div>
+                </div>
+            {% endfor %}
+            <div class="row">
+                <div class="col-sm">Guardians</div>
+            </div>
+            {% for character in object.guardian.all %}
+                <div class="row">
+                    <div class="col-sm">{{ character.name }}</div>
+                </div>
+            {% endfor %}
+            <div class="row">
+                <div class="col-sm">Teachers</div>
+            </div>
+            {% for character in object.teacher.all %}
+                <div class="row">
+                    <div class="col-sm">
+                        {{ character.name }}
+                    </div>
+                </div>
+            {% endfor %}
+            <div class="row">
+                <div class="col-sm">
+                </div>
+                <div class="col-sm">
+                </div>
+                <div class="col-sm">
+                    Reality Zone
+                </div>
+                <div class="col-sm">
+                    {{ object.reality_zone }} ({{ object.reality_zone_rating }})
+                </div>
+                <div class="col-sm">
+                </div>
+                <div class="col-sm">
+                </div>
             </div>
             <div class="row">
-                <div class="col-sm">Allies</div>
-                <div class="col-sm">{{ object.allies }}</div>
-                <div class="col-sm">Backup</div>
-                <div class="col-sm">{{ object.backup }}</div>
-                <div class="col-sm">Cult</div>
+                <div class="col-sm">
+                    Gauntlet
+                </div>
+                <div class="col-sm">
+                    {{ object.gauntlet }}
+                </div>
+                <div class="col-sm">
+                    Shroud
+                </div>
+                <div class="col-sm">
+                    {{ object.shroud }}
+                </div>
+                <div class="col-sm">
+                    Dimension Barrier
+                </div>
+                <div class="col-sm">
+                    {{ object.dimension_barrier }}
+                </div>
+            </div>
+            <div class="row">
+                <div class="col-sm">
+                    Allies
+                </div>
+                <div class="col-sm">
+                    {{ object.allies }}
+                </div>
+                <div class="col-sm">
+                    Backup
+                </div>
+                <div class="col-sm">
+                    {{ object.backup }}
+                </div>
+                <div class="col-sm">
+                    Cult
+                </div>
                 <div class="col-sm">
                     {{ object.cult }}
                 </div>
@@ -108,6 +178,21 @@
                     {{ object.requisitions }}
                 </div>
             </div>
+            <h2 class="col-sm mtas_heading">
+                Cabals
+            </h2>
+            {% for cabal in object.cabals.all %}
+                <div class="row">
+                    <a class="col-sm" href={{ cabal.get_absolute_url }}><h3 class="mtas_heading">{{ cabal.name }}</h3></a>
+                </div>
+                {% for member in cabal.members.all %}
+                    <div class="row">
+                        <div class="col-sm">
+                            <a class="col-sm" href={{ member.get_absolute_url }}>{{ member.name }}</a>
+                        </div>
+                    </div>
+                {% endfor %}
+            {% endfor %}
             <h2 class="col-sm mtas_heading">
                 Nodes
             </h2>
